requirements.txt
Django==2.0.3
djangorestframework==3.7.7
feedparser==5.2.1
pytz==2018.3
